-created currently reading book screen with the progress info -implemented "start to read" in the book screen -created the comments in progress screen -fixed some errors

+++ app/src/main/java/com/ayoprez/sobuu/shared/models/api_models/currently_reading_api/Extras.kt
@@ -6,7 +6,7 @@ import com.squareup.moshi.JsonClass
 data class Extras(
     val commentsInProgress: List<CommentsInProgress>,
     val currentPercentage: Int? = null,
-    val currentProgress: Int,
+    val currentProgress: Double,
     val currentPage: Int? = null,
     val startedToRead: StartedToRead
 )
